test(omo-senpi): split recall wiring tests by subject
Plan: .omo/plans/memorian-tool-boundary.md

Ultraworked with omo (https://github.com/code-yeongyu/oh-my-openagent)
